2025/python/pyproject.toml
[project]
name = "python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "gitpython>=3.1.45",
    "numpy>=2.3.5",
    "requests>=2.32.5",
]
